requirements.txt
Pillow
Werkzeug
aiofiles
aioredis
debugpy
fastapi
flash
jinja2
pydantic
python-multipart
uvicorn
